ipc (message queue) implementated
syscall and blocking read still unimplemented

+++ Userland/help/stdlib/stdin.c
@@ -5,7 +5,7 @@ extern uint64_t systemCall(uint64_t eax, uint64_t rbx, uint64_t rcx, uint64_t rd
 
 char * gets(char * s)
 {
-  systemCall(0x03, 0x01, s, 0, 0, 0);
+  systemCall(0x03, 0x01, (uint64_t)s, 0, 0, 0);
   while(*s!='\n') { s++; }
   *s = '\0';
   return s;
